# SlottedArray supports 16kb