@Scooletz 2024-09-11
Not the greatest results from benchmarks. Will take a look later on