@Scooletz 2024-09-24
This made it too slow to be applicable.